stash: Remove the stash
