+++ .config/nextest.toml
@@ -17,10 +17,6 @@ filter = "package(mz-environmentd)"
 threads-required = 8
 slow-timeout = { period = "120s", terminate-after = 2 }
 
-[[profile.default.overrides]]
-filter = "package(mz-stash) and (test(proptest_stash_migrate_json_to_proto) or test(test_stash_batch_large_number_updates))"
-slow-timeout = { period = "120s", terminate-after = 2 }
-
 [profile.ci]
 junit = { path = "junit_cargo-test.xml" }
 fail-fast = false
