+++ misc/images/jobs/.gitignore
@@ -1,3 +1,2 @@
 /persistcli
-/mz-stash-debug
 /mz-catalog-debug
