+++ misc/images/jobs/Dockerfile
@@ -10,7 +10,6 @@
 MZFROM prod-base
 
 COPY persistcli /usr/local/bin/persistcli
-COPY mz-stash-debug /usr/local/bin/mz-stash-debug
 COPY mz-catalog-debug /usr/local/bin/mz-catalog-debug
 
 # TODO: the intention is for this image to eventually serve as the one-off
