+++ src/adapter/src/util.rs
@@ -358,7 +358,6 @@ impl<T> ShouldHalt for StorageError<T> {
             | StorageError::DataflowError(_)
             | StorageError::InvalidAlter { .. }
             | StorageError::ShuttingDown(_) => false,
-            StorageError::IOError(e) => e.is_unrecoverable(),
         }
     }
 }
