+++ src/catalog/src/durable/error.rs
@@ -12,7 +12,6 @@ use std::fmt::Debug;
 use mz_proto::TryFromProtoError;
 use mz_repr::Timestamp;
 use mz_sql::catalog::CatalogError as SqlCatalogError;
-use mz_stash_types::StashError;
 use mz_storage_types::controller::StorageError;
 
 #[derive(Debug, thiserror::Error)]
@@ -69,9 +68,6 @@ pub enum DurableCatalogError {
     /// A programming error occurred during a [`mz_storage_client::controller::StorageTxn`].
     #[error(transparent)]
     Storage(StorageError<Timestamp>),
-    // TODO: remove this with migration from storage stash to persist catalog.
-    #[error(transparent)]
-    Stash(StashError),
 }
 
 impl DurableCatalogError {
@@ -81,8 +77,7 @@ impl DurableCatalogError {
             DurableCatalogError::Fence(_)
             | DurableCatalogError::IncompatibleDataVersion { .. }
             | DurableCatalogError::IncompatiblePersistVersion { .. }
-            | DurableCatalogError::Proto(_)
-            | DurableCatalogError::Stash(_) => true,
+            | DurableCatalogError::Proto(_) => true,
             DurableCatalogError::Uninitialized
             | DurableCatalogError::NotWritable(_)
             | DurableCatalogError::DuplicateKey
@@ -118,11 +113,5 @@ impl From<StorageError<Timestamp>> for DurableCatalogError {
 impl From<TryFromProtoError> for DurableCatalogError {
     fn from(e: TryFromProtoError) -> Self {
         DurableCatalogError::Proto(e)
-    }
-}
-
-impl From<StashError> for DurableCatalogError {
-    fn from(e: StashError) -> Self {
-        DurableCatalogError::Stash(e)
     }
 }
