+++ src/controller/src/lib.rs
@@ -50,7 +50,6 @@ use mz_persist_types::Codec64;
 use mz_proto::RustType;
 use mz_repr::{GlobalId, TimestampManipulation};
 use mz_service::secrets::SecretsReaderCliArgs;
-use mz_stash_types::metrics::Metrics as StashMetrics;
 use mz_storage_client::client::{
     ProtoStorageCommand, ProtoStorageResponse, StorageCommand, StorageResponse,
 };
@@ -93,8 +92,6 @@ pub struct ControllerConfig {
     pub init_container_image: Option<String>,
     /// The now function to advance the controller's introspection collections.
     pub now: NowFn,
-    /// The process-wide stash metrics.
-    pub stash_metrics: Arc<StashMetrics>,
     /// The metrics registry.
     pub metrics_registry: MetricsRegistry,
     /// The URL for Persist PubSub.
