+++ src/proto/Cargo.toml
@@ -12,6 +12,7 @@ workspace = true
 
 [dependencies]
 anyhow = "1.0.66"
+bytes = "1.3.0"
 chrono = { version = "0.4.35", default-features = false, features = ["serde", "std"], optional = true }
 chrono-tz = { version = "0.8.1", features = ["serde", "case-insensitive"], optional = true }
 globset = "0.4.9"
@@ -19,7 +20,7 @@ http = "0.2.8"
 mz-ore = { path = "../ore", default-features = false, features = ["proptest", "test"] }
 num = "0.4.0"
 proptest = { version = "1.0.0", default-features = false, features = ["std"] }
-proptest-derive = { version = "0.3.0", features = ["boxed_union"]}
+proptest-derive = { version = "0.3.0", features = ["boxed_union"] }
 prost = { version = "0.11.3", features = ["no-recursion-limit"] }
 regex = "1.7.0"
 serde = { version = "1.0.152", features = ["derive", "rc"] }
