--- src/stash-debug/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "mz-stash-debug"
-description = "Durable metadata storage."
-version = "0.95.0-dev"
-edition.workspace = true
-rust-version.workspace = true
-publish = false
-
-[lints]
-workspace = true
-
-[dependencies]
-anyhow = "1.0.66"
-clap = { version = "3.2.24", features = ["derive", "env"] }
-mz-build-info = { path = "../build-info" }
-mz-ore = { path = "../ore" }
-mz-tls-util = { path = "../tls-util" }
-mz-stash = { path = "../stash" }
-mz-storage-controller = { path = "../storage-controller" }
-once_cell = "1.16.0"
-serde_json = "1.0.89"
-tokio = "1.32.0"
-tokio-postgres = { version = "0.7.8", features = ["with-serde_json-1"] }
-tracing-subscriber = { version = "0.3.16", default-features = false, features = ["env-filter", "fmt"] }
-workspace-hack = { version = "0.0.0", path = "../workspace-hack" }
-
-[package.metadata.cargo-udeps.ignore]
-normal = ["workspace-hack"]
