--- src/stash-debug/README.md
@@ -1,17 +1,0 @@
-# `stash-debug`
-
-This tool allows us to debug an individual customer's Stash. It connects directly to CockroachDB so
-you will need a Cloud Admin to help you.
-
-> **Note**: If you plan on running `stash-debug` more than once or on a very large Stash, we'd
-recommend compiling it in release mode, e.g. `cargo run --release -- <args>`.
-
-### `upgrade-check`
-
-To use the `upgrade-check` command you'll need to provide a mapping from cluster replica sizes to
-resource specification. To get the latest mapping check the [Pulumi config in the Cloud Repo](https://github.com/MaterializeInc/cloud/blob/main/Pulumi.production.yaml), search for "cluster_replica_sizes".
-
-A nice oneliner for converting from YAML to JSON is:
-```
-python -c 'import sys,json,yaml; print(json.dumps(yaml.safe_load(sys.stdin.read())))'
-```
