+++ src/stash/Cargo.toml
@@ -21,7 +21,6 @@ fail = { version = "0.5.1", features = ["failpoints"] }
 futures = "0.3.25"
 mz-ore = { path = "../ore", features = ["metrics", "network", "async", "test"] }
 mz-proto = { path = "../proto" }
-mz-stash-types = { path = "../stash-types" }
 mz-tls-util = { path = "../tls-util" }
 paste = "1.0.11"
 postgres-openssl = { version = "0.5.0" }
