+++ src/testdrive/Cargo.toml
@@ -33,7 +33,7 @@ junit-report = "0.8.3"
 once_cell = "1.16.0"
 maplit = "1.0.2"
 md-5 = "0.10.5"
-mysql_async = { version = "0.34.1",  default-features = false, features = ["minimal"] }
+mysql_async = { version = "0.34.1", default-features = false, features = ["minimal"] }
 mz-avro = { path = "../avro", features = ["snappy"] }
 mz-aws-util = { path = "../aws-util", features = ["s3"] }
 mz-build-info = { path = "../build-info" }
@@ -51,7 +51,6 @@ mz-repr = { path = "../repr" }
 mz-sql = { path = "../sql" }
 mz-sql-parser = { path = "../sql-parser" }
 mz-storage-types = { path = "../storage-types" }
-mz-stash = { path = "../stash" }
 mz-tls-util = { path = "../tls-util" }
 postgres_array = { version = "0.11.0" }
 prost = { version = "0.11.3", features = ["no-recursion-limit"] }
